created budget component services

+++ src/app/shared/components/category-card/category-card.component.ts
@@ -5,12 +5,19 @@ import { NotificationService } from '../../services/general/notification.service
 import { HttpStatusCode } from '../../../constents/http-status-code';
 import { Router } from '@angular/router';
 import { EditCategoryComponent } from 'src/app/category/edit-category/edit-category.component';
+import { ConfirmDeleteComponent } from '../confirm-delete/confirm-delete.component';
+import { yearsPerPage } from '@angular/material/datepicker';
+import { CommonService } from '../../services/common.service';
+import { nextTick } from 'process';
+import { ResourceLoader } from '@angular/compiler';
+
 @Component({
   selector: 'app-category-card',
   templateUrl: './category-card.component.html',
   styleUrls: ['./category-card.component.scss'],
 })
 export class CategoryCardComponent implements OnInit {
+  data;
   @Output() isDelete = new EventEmitter<boolean>();
   @Input()
   category: {
@@ -19,8 +26,11 @@ export class CategoryCardComponent implements OnInit {
     isActive: boolean;
   };
   constructor(
+    
     public dialog: MatDialog,
+    private router:Router,
     private categoryService: CategoryService,
+    private commonService:CommonService,
     private notficationService: NotificationService
   ) {}
 
@@ -37,13 +47,44 @@ export class CategoryCardComponent implements OnInit {
     });
   }
   deleteCategory(category) {
-    this.categoryService.deleteCategory(category).subscribe((res: any) => {
-      if (res.status == HttpStatusCode.OK) {
-        this.notficationService.showSuccessMessage(
-          'Category Deleted Sucessfully'
-        );
-        this.isDelete.emit(true);
-      }
+    
+    const dialogRef = this.dialog.open(ConfirmDeleteComponent,{
+      width:'400px',
+      data:{category}
     });
-  }
+
+
+     this.commonService.subsVar = this.commonService
+     .invokeConfirmComponent.subscribe((flag) =>{
+       if(flag == 1 ){
+        this.categoryService.deleteCategory(category).subscribe((res: any) => {
+          this.data = res;
+          
+          if (this.data.status == HttpStatusCode.OK) {
+            
+            this.notficationService.showSuccessMessage(
+              'Category Deleted Sucessfully'
+            );
+            
+            this.isDelete.emit(true);
+            this.dialog.closeAll();
+            location.reload();
+          }
+        });
+       }
+       
+     })
+    
 }
+
+    // this.categoryService.deleteCategory(category).subscribe((res: any) => {
+    //   if (res.status == HttpStatusCode.OK) {
+        
+    //     this.notficationService.showSuccessMessage(
+    //       'Category Deleted Sucessfully'
+    //     );
+    //     this.isDelete.emit(true);
+    //   }
+    // });
+
+  }
